Add files via upload
Sequential search algorithm in C. It searches through each element of an array until it finds the element which the user inputted initially. If not found, the program lets the user know that their element is not present in the array. If found, the program tells the user at what position the element is located within the array.